unify types
