+++ examples/docker-compose.ebpf.yml
@@ -19,7 +19,7 @@ services:
       - HUGINN_EBPF_DST_PORT=7000
       - HUGINN_EBPF_PIN_PATH=/sys/fs/bpf/huginn
       - HUGINN_EBPF_SYN_MAP_MAX_ENTRIES=8192
-      - HUGINN_EBPF_XDP_MODE=${XDP_MODE:-skb}
+      - HUGINN_EBPF_CAPTURE=${HUGINN_EBPF_CAPTURE:-xdp-skb}
       - HUGINN_EBPF_METRICS_ADDR=0.0.0.0
       - HUGINN_EBPF_METRICS_PORT=9091
     healthcheck:
